1. 수수료 소수점 적용  - 소수점 두자리 버그 수정  - RateVO.lrIdx Byte -> Integer로 수정

+++ src/main/java/com/tagosolution/service/model/RateVO.java
@@ -9,7 +9,7 @@ public class RateVO implements GeneralModel {
 	 */
 	private static final long serialVersionUID = -1L;
 
-	private Byte lrIdx;
+	private Integer lrIdx;
 
 	private Integer runTime;
 
@@ -25,11 +25,11 @@ public class RateVO implements GeneralModel {
 
     private Float lineRateT;
 
-    public Byte getLrIdx() {
+    public Integer getLrIdx() {
 		return lrIdx;
 	}
 
-	public void setLrIdx(Byte lrIdx) {
+	public void setLrIdx(Integer lrIdx) {
 		this.lrIdx = lrIdx;
 	}
